bus_safety_ai/requirements.txt
numpy
opencv-python
mediapipe
tensorflow
pyserial
